backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.18.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
dramatiq[redis]==1.18.0
redis==5.2.1
python-dotenv==1.0.1
google-generativeai==0.8.5
httpx==0.28.1
pydantic==2.11.7
websockets==15.0.1
tavily-python==0.5.4
daytona-sdk==0.21.0